+++ src/main/java/com/reason/ide/files/MlgFileViewProviderFactory.java
@@ -1,0 +1,13 @@
+package com.reason.ide.files;
+
+import com.intellij.lang.*;
+import com.intellij.openapi.vfs.*;
+import com.intellij.psi.*;
+import org.jetbrains.annotations.*;
+
+public class MlgFileViewProviderFactory implements FileViewProviderFactory {
+    @Override
+    public @NotNull FileViewProvider createFileViewProvider(@NotNull VirtualFile file, Language language, @NotNull PsiManager manager, boolean eventSystemEnabled) {
+        return new MlgFileViewProvider(manager, file, eventSystemEnabled);
+    }
+}