+++ src/main/java/com/reason/lang/ocamlgrammar/OclGrammarLanguage.java
@@ -1,0 +1,11 @@
+package com.reason.lang.ocamlgrammar;
+
+import com.intellij.lang.*;
+
+public class OclGrammarLanguage extends Language {
+    public static final OclGrammarLanguage INSTANCE = new OclGrammarLanguage();
+
+    private OclGrammarLanguage() {
+        super("Mlg");
+    }
+}